[refactor] Extract the space-leave flow from the entrypoint
space:leave was a 223-line state machine living in the file whose job is
wiring, and that is why the reusable half of the flow was written separately
in space.js and drifted: both encoded the same five-step teardown order
independently. It moves to worker/ipc/space-leave.js behind a
registerSpaceLeave(ipc, deps), and the order both callers share now lives in
spaces/leave-flow.js.

The handler body moves rather than being rewritten; the ten leave flow tests
pass unedited. One behaviour does change: the mount, share and foreign steps
were wrapped in a single try, so a failure in the first skipped the rest —
they are isolated now, matching what the boot pass already did.
